Support PHP 8
