+++ .php_cs.dist
@@ -13,7 +13,7 @@ HEADER;
 
 $finder = PhpCsFixer\Finder::create()
     ->in(__DIR__)
-    ->exclude('features/var');
+    ->exclude(['features', 'flex']);
 
 return PhpCsFixer\Config::create()
     ->setRiskyAllowed(true)
