+++ composer.json
@@ -26,7 +26,7 @@
         "php": "^7.2",
         "ext-SimpleXML": "*",
         "docusign/esign-client": "^3.0",
-        "lcobucci/jwt": "^3.3.1",
+        "lcobucci/jwt": "^3.3.1 || ^4.0-dev",
         "league/flysystem": "^1.0.8",
         "psr/log": "^1.1",
         "symfony/config": "^4.4 || ^5.0",
@@ -43,24 +43,23 @@
         "webmozart/assert": "^1.5"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.16.2",
+        "doctrine/annotations": "^1.11",
         "league/flysystem-bundle": "^1.2",
-        "nyholm/symfony-bundle-test": "^1.6.1",
-        "phpspec/prophecy": "^1.9",
+        "nyholm/symfony-bundle-test": "dev-master",
         "phpstan/phpstan": "^0.12.18",
-        "phpunit/phpunit": "^7.5.2 || ^8.0",
         "psr/event-dispatcher": "^1.0",
+        "symfony/console": "^4.4 || ^5.0",
         "symfony/css-selector": "^4.4 || ^5.0",
         "symfony/debug": "^4.4 || ^5.0",
         "symfony/dotenv": "^4.4 || ^5.0",
         "symfony/monolog-bundle": "^3.5",
         "symfony/panther": "^0.6.1",
+        "symfony/phpunit-bridge": "^4.4 || ^5.0",
         "symfony/polyfill-php72": "^1.9",
         "symfony/process": "^4.4 || ^5.0",
         "symfony/profiler-pack": "^1.0",
         "symfony/security-bundle": "^4.4 || ^5.0",
-        "symfony/twig-bundle": "^4.4 || ^5.0",
-        "symfony/web-server-bundle": "^4.4 || ^5.0"
+        "symfony/twig-bundle": "^4.4 || ^5.0"
     },
     "suggest": {
         "league/flysystem-bundle": "For Symfony >4.2 - Allow the use of the official adapter system instead of the compat layer.",
@@ -74,8 +73,17 @@
     "autoload-dev": {
         "psr-4": {
             "DocusignBundle\\Tests\\": "tests/",
-            "DocusignBundle\\E2e\\": "features/"
+            "DocusignBundle\\E2e\\": "features/src/",
+            "PHPUnit\\": "vendor/bin/.phpunit/phpunit/src"
         }
+    },
+    "scripts": {
+        "post-install-cmd": [
+            "vendor/bin/simple-phpunit --version"
+        ],
+        "post-update-cmd": [
+            "vendor/bin/simple-phpunit --version"
+        ]
     },
     "config": {
         "sort-packages": true
