--- features/TestBundle/DependencyInjection/TestExtension.php
@@ -1,35 +1,0 @@
-<?php
-
-/*
- * This file is part of the DocusignBundle.
- *
- * (c) Grégoire Hébert <gregoire@les-tilleuls.coop>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace DocusignBundle\E2e\TestBundle\DependencyInjection;
-
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-
-/**
- * @author Vincent Chalamon <vincentchalamon@gmail.com>
- */
-final class TestExtension extends Extension
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function load(array $configs, ContainerBuilder $container): void
-    {
-        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('services.yaml');
-        $loader->load('override.yaml');
-    }
-}
