--- features/TestBundle/Resources/config/override.yaml
@@ -1,4 +1,0 @@
-services:
-  test.docusign.token_encoder.default:
-    parent: docusign.token_encoder.default
-    public: true
