--- features/TestBundle/Resources/config/services.yaml
@@ -1,22 +1,0 @@
-# Those parameters are required to retrieve secrets in e2e tests
-parameters:
-  docusign.email: '%env(DOCUSIGN_EMAIL)%'
-  docusign.password: '%env(DOCUSIGN_PASSWORD)%'
-
-services:
-  _defaults:
-    autowire: true
-    autoconfigure: true
-    public: false
-    bind:
-      $kernelProjectDir: '%kernel.project_dir%'
-
-  Symfony\Component\HttpFoundation\Session\Session: '@session'
-
-  DocusignBundle\E2e\TestBundle\:
-    resource: '../../*'
-    exclude: '../../{DependencyInjection,TestBundle.php}'
-
-  DocusignBundle\E2e\TestBundle\Controller\:
-    resource: '../../Controller/*'
-    tags: ['controller.service_arguments']
