--- features/TestBundle/Resources/views/base.html.twig
@@ -1,42 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>{{ block('title')}}</title>
-    <style>
-        body{background:#F5F5F5;font:18px/1.5 sans-serif;}
-        h1{line-height:1.2;margin:0 0 1em;font-size:36px;float:left;max-width:650px;}
-        p{margin:0 0 1em 0;}
-        a{color:#0000F0;}
-        a:hover{text-decoration:none;}
-        code{background:#F5F5F5;max-width:100px;padding:2px 6px;word-wrap:break-word;}
-        #wrapper{background:#FFF;margin:1em auto;max-width:800px;width:95%;}
-        #container{padding:2em;}
-        #content{clear:both;}
-        #footer p{color:#777777;background:#FEFFEA;padding:10px;font-size:14px;margin-bottom:0;}
-        #footer p a{color:#777777;text-decoration:none;}
-        #footer p a:hover{color:#333333;}
-        .alert{clear:both;color:#155724;background:#d4edda;border:thin solid #c3e6cb;border-radius:.25rem;padding:10px;font-size:14px;}
-        #logout,#login{float:right;margin-top:8px;}
-    </style>
-    {% block stylesheets %}{% endblock %}
-</head>
-<body>
-<div id="wrapper">
-    <div id="container">
-        <h1>{% block title %}{% endblock %}</h1>
-        {% if app.user %}
-            <a href="{{ path('logout') }}" id="logout">Log out</a>
-        {% elseif 'login' != app.request.get('_route') %}
-            <a href="{{ path('login') }}" id="login">Log in</a>
-        {% endif %}
-        {% for flashMessage in app.session.flashbag.get('success') %}
-            <div class="alert">
-                {{ flashMessage|raw }}
-            </div>
-        {% endfor %}
-        <div id="content">{% block content %}{% endblock %}</div>
-    </div>
-    <div id="footer">{% block footer %}{% endblock %}</div>
-</div>
-</body>
-</html>
