--- features/TestBundle/Resources/views/homepage.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Welcome to the DocuSign TestBundle{% endblock %}
-
-{% block content %}
-    <ul>
-        <li><a href="{{ path('embedded') }}">Embedded mode (JWT auth)</a></li>
-        <li><a href="{{ path('remote') }}">Remote mode (JWT auth)</a></li>
-        <li><a href="{{ path('embedded_auth_code') }}">Embedded mode (AuthorizationCode auth)</a></li>
-        <li><a href="{{ path('remote_auth_code') }}">Remote mode (AuthorizationCode auth)</a></li>
-        <li><a href="{{ path('clickwrap') }}">Clickwrap mode</a></li>
-    </ul>
-{% endblock %}
