--- features/TestBundle/Resources/views/list.html.twig
@@ -1,15 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Documents list ({{ mode }} mode - {{ auth }} auth){% endblock %}
-
-{% block content %}
-    <ul>
-        {% for name, path in documents %}
-            <li><a href="{{ path(route_name, {'path': path, 'signerName': 'John DOE'}) }}">{{ name }}</a></li>
-        {% endfor %}
-    </ul>
-{% endblock %}
-
-{% block footer %}
-    <p><a href="{{ path('homepage') }}">&#8592; Back to the homepage</a></p>
-{% endblock %}
