--- features/TestBundle/TestBundle.php
@@ -1,23 +1,0 @@
-<?php
-
-/*
- * This file is part of the DocusignBundle.
- *
- * (c) Grégoire Hébert <gregoire@les-tilleuls.coop>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace DocusignBundle\E2e\TestBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-/**
- * @author Vincent Chalamon <vincentchalamon@gmail.com>
- */
-final class TestBundle extends Bundle
-{
-}
