+++ features/config/override.yaml
@@ -1,0 +1,4 @@
+services:
+  test.docusign.token_encoder.default:
+    parent: docusign.token_encoder.default
+    public: true
