+++ features/config/routes/annotations.yaml
@@ -1,0 +1,3 @@
+controllers:
+  resource: ../../src/Controller/
+  type: annotation
