+++ features/config/routes/debug.yaml
@@ -1,0 +1,11 @@
+_errors:
+  resource: '@FrameworkBundle/Resources/config/routing/errors.xml'
+  prefix: /_error
+
+web_profiler_wdt:
+  resource: '@WebProfilerBundle/Resources/config/routing/wdt.xml'
+  prefix: /_wdt
+
+web_profiler_profiler:
+  resource: '@WebProfilerBundle/Resources/config/routing/profiler.xml'
+  prefix: /_profiler
