+++ features/config/routes/docusign.yaml
@@ -1,0 +1,3 @@
+docusign:
+  resource: .
+  type: docusign
