+++ features/config/services.yaml
@@ -1,0 +1,21 @@
+# Those parameters are required to retrieve secrets in e2e tests
+parameters:
+  docusign.email: '%env(DOCUSIGN_EMAIL)%'
+  docusign.password: '%env(DOCUSIGN_PASSWORD)%'
+
+services:
+  _defaults:
+    autowire: true
+    autoconfigure: true
+    public: false
+    bind:
+      $kernelProjectDir: '%kernel.project_dir%'
+
+  Symfony\Component\HttpFoundation\Session\Session: '@session'
+
+  DocusignBundle\E2e\:
+    resource: '../src/*'
+
+  DocusignBundle\E2e\Controller\:
+    resource: '../src/Controller/*'
+    tags: ['controller.service_arguments']
