+++ features/console
@@ -1,6 +1,7 @@
 #!/usr/bin/env php
 <?php
 
+use DocusignBundle\E2e\Kernel;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Debug\Debug;
