+++ features/public/index.php
@@ -11,6 +11,7 @@
 
 declare(strict_types=1);
 
+use DocusignBundle\E2e\Kernel;
 use Symfony\Component\Debug\Debug;
 use Symfony\Component\HttpFoundation\Request;
